requirements.txt
fastapi==0.95.1
uvicorn==0.17.5
SQLAlchemy==2.0.25
PyMySQL==1.0.2
pydantic==1.9.0
pytest==7.0.1
requests==2.28.2
wheel==0.41.2
python-multipart==0.0.5
python-dotenv==0.20.0
mangum==0.15.0
aiomysql==0.1.0
nest-asyncio==1.5.5
itsdangerous==2.1.2
Jinja2==3.1.2
passlib==1.7.4
python-dateutil==2.8.2
us==2.0.2
pytz==2022.7.1
aiohttp==3.8.4
frozendict==2.3.8
async-lru==2.0.3
PyJWT==2.7.0
alembic==1.13.1
mysqlclient==2.2.3
